\name{hideFiles}

\alias{hideFiles}

\title{Ascunde fisiere si directoare din build}

\description{
Ascunde fisiere si directoare din build
}

\usage{
hideFiles(project = NULL, gitdir = NULL)
}

\arguments{
  \item{project}{Numele directorului / proiectului de interes}
  \item{gitdir}{Directorul unde sunt gazduite proiectele}
}


\author{
Adrian Dusa
}

\keyword{functions}
