\name{makeSQL}

\alias{makeSQL}

\title{Creaza fisier SQL}

\description{
Creaza fisier .sql pentru definirea variabilelor din tabela SQL
}

\usage{
makeSQL(excel, sheet = 1, tabela = NULL, dir = NULL)
}

\arguments{
  \item{excel}{Calea catre fisierul .ods de unde se citesc informatiile}
  \item{sheet}{Numele sau numarul foii care sa fie citita}
  \item{tabela}{Numele tabelei SQL care sa fie creata}
  \item{dir}{Optional, numele directorului unde sa fie creat fisierul}
}


\author{
Adrian Dusa
}

\keyword{functions}
