\name{makeVars}

\alias{makeVars}

\title{Creaza fisierul .ts pentru variabile}

\description{
Fisierul js este un modul exportat care contine componentele questions, questionsOrder si activators.
}

\usage{
makeVars(excel, sheet = "", destination = NULL, ...)
}

\arguments{
  \item{excel}{Calea catre fisierul .ods de unde se citesc informatiile}
  \item{sheet}{Numele foii care sa fie citita}
  \item{destination}{Calea catre directorul unde se scriu informatiile in fisierul .ts}
  \item{...}{Alte argumente}
}


\author{
Adrian Dusa
}

\keyword{functions}
