\name{pocufiles}

\alias{pocufiles}

\title{Split in fisiere sursa}

\description{
Sparge un document Excel agregat, in mai multe fisiere sursa.
}

\usage{
pocufiles(datacsv, sheet = NULL, template = "", dirname = "")
}

\arguments{
  \item{datacsv}{Calea catre fisierul CSV cu datele brute}
  \item{sheet}{Numarul foii din fisierul Excel care sa fie citita}
  \item{template}{Calea catre fisierul template}
  \item{dirname}{Calea catre directorul unde se vor crea datele}
}


\author{
Adrian Dusa
}

\keyword{functions}
