\name{splitHtml}

\alias{splitHtml}

\title{Genereaza fisiere html pe sectiuni}

\description{
Genereaza fisiere html pe sectiuni
}

\usage{
splitHtml(html, dir, ...)
}

\arguments{
  \item{html}{Calea catre fisierul .html de unde se citesc informatiile}
  \item{dir}{Calea catre directorul unde se scriu fisierele}
  \item{...}{Alte argumente}
}


\author{
Adrian Dusa
}

\keyword{functions}
