\name{WB-package}

\alias{WB-package}

\alias{WB}

\docType{package}

\title{
WB: World Bank
}

\description{
Fisiere utile pentru chestionarele WB
}

\details{
    \tabular{ll}{
    Package: \tab WB\cr
    Type:    \tab Package\cr
    Version: \tab 0.1\cr
    Date:    \tab 2020-10-23\cr
    License: \tab GPL (>= 2)\cr
    }
}

\author{

\bold{Authors}:\cr
Adrian Dusa\cr
\email{dusa.adrian@gmail.com}

\bold{Maintainer}:\cr
Adrian Dusa
}

\keyword{package}

