\name{hideFiles}

\alias{hideFiles}

\title{Ascunde fisiere si directoare din build}

\description{
Ascunde fisiere si directoare din build
}

\usage{
hideFiles(gitdir = NULL, project = NULL)
}

\arguments{
  \item{gitdir}{Directorul unde sunt gazduite proiectele}
  \item{sheet}{Numele directorului / proiectului de interes}
}


\author{
Adrian Dusa
}

\keyword{functions}
