\name{makeVars}

\alias{makeVars}

\title{Creaza fisierul .js pentru variabile}

\description{
Fisierul js este un modul exportat care contine componentele questions, questionsOrder si activators.
}

\usage{
makeVars(ods, js)
}

\arguments{
  \item{ods}{Calea catre fisierul .ods de unde se citesc informatiile}
  \item{js}{Calea catre fisierul .js unde se scriu informatiile}
}


\author{
Adrian Dusa
}

\keyword{functions}
